api/requirements.txt
tensorflow==2.17.1
tensorflow-serving-api==2.17.1
fastapi
uvicorn
python-multipart
pillow
matplotlib
numpy
